Usage of new elcodi related purchasable service
* Using related products as subrequests, using new service
* Deprecated ProductCollectionProvider in favor of new elcodi/elcodi implementation
* Fixed some finish translations (only format)

Also : Removed logic from listener

* This change is because of new final events
* Placed this logic into a service
* Fixed tests (listeners shouln't have tests. Now service is tested)
