+++ composer.json
@@ -74,7 +74,7 @@
         "mmoreram/cache-flush-bundle": "0.1.2",
         "mmoreram/http-headers-bundle": "1.0.0",
 
-        "elcodi/elcodi": "^1.0"
+        "elcodi/elcodi": "^1.0.6"
     },
     "replace": {
         "elcodi-templates/template": "self.version",
