+++ src/Elcodi/Plugin/CustomShippingBundle/Resources/config/services.yml
@@ -7,3 +7,13 @@ services:
         parent: elcodi.abstract_plugin
         arguments:
             - "Elcodi\\Plugin\\CustomShippingBundle\\ElcodiCustomShippingBundle"
+
+    #
+    # Providers
+    #
+    elcodi_plugin.custom_shipping.provider.shipping_ranges:
+        class: Elcodi\Plugin\CustomShippingBundle\Provider\ShippingRangesProvider
+        arguments:
+            - @elcodi.repository.carrier
+            - @elcodi.converter.currency
+            - @elcodi.matcher.zone
