+++ src/Elcodi/Plugin/StoreTemplateBundle/Resources/views/Pages/cart-view.html.twig
@@ -56,5 +56,7 @@
 {% endblock %}
 
 {% block prefooter %}
-    {% include "@StoreTemplate/Modules/_product-related.html.twig" with { 'products' : related_products} %}
+    {% if cart.totalItemNumber > 0 %}
+        {{ render(url('store_cart_related')) }}
+    {% endif %}
 {% endblock prefooter %}
