+++ src/Elcodi/Store/ProductBundle/Services/StoreCategoryTree.php
@@ -22,6 +22,9 @@ use Elcodi\Component\Language\Entity\Interfaces\LocaleInterface;
 use Elcodi\Component\Product\Entity\Category;
 use Elcodi\Component\Product\Services\CategoryTree;
 
+/**
+ * Class StoreCategoryTree
+ */
 class StoreCategoryTree extends AbstractCacheWrapper
 {
     /**
